update for stageII bootloader
Refinements for the packet encoder as well as the communication handling
have been made. The so called 'stage II' bootloader is named because it
is the second bootloader that can be loaded in code. The first is passed
over unless the word at 0xfffe is changed. The stage II bootloader
appears to have a 512 byte limit for flashing.