v2.12
I made changes so that the pseudo-real-time tuning part works. This code
reads in a binary file and loads it to the ECU. When the binary file
changes, it loads only the changed part to the ECU. I made a bunch of
ancillary changes to the code along my way to work out bugs in the
real-time part, so I may well have introduced new failure modes. For
instance, I adjusted functions in the main 'serial packet' class. This
may affect reflashing packets. I also changed byte logging to be in hex
instead of decimal. This affects reflashing as well.